Refactoring(InheritanceModel, MavenLauncher)
